requirements.txt
accelerate>=0.16.0,<1
click>=8.0.4,<9
datasets>=2.10.0,<3
deepspeed>=0.8.3,<0.9
# torch==2.0.1
# torchvision==0.15.2
transformers[torch]>=4.28.1,<5
langchain>=0.0.139
albumentations==0.4.3
opencv-python==4.8.0.76
pudb==2019.2
pydantic==1.10.9
imageio==2.9.0
imageio-ffmpeg==0.4.2
pytorch-lightning==1.0.8
omegaconf==2.0.0
test-tube>=0.7.5
streamlit>=0.73.1
einops==0.3.0
more-itertools>=8.0.0
transformers==4.30.2
matplotlib==3.7.2
taming-transformers
# setuptools==59.5.0